requirements.txt
django==1.10.8
djangorestframework==3.7.7
flake8==3.5.0
Pillow==5.0.0
